// query-node/types/generated/2001/types-lookup.ts
export interface PalletForumPrivilegedActor extends Enum {
  readonly isLead: boolean;
  readonly isModerator: boolean;
  readonly asModerator: u64;
  readonly type: "Lead" | "Moderator";
}

/** @name PalletForumExtendedPostIdObject (80) */
export interface PalletForumExtendedPostIdObject extends Struct {
  readonly categoryId: u64;
  readonly threadId: u64;
  readonly postId: u64;
}